chart

+++ db/migrate/20160425204244_create_strokes.rb
@@ -1,8 +1,8 @@
 class CreateStrokes < ActiveRecord::Migration
   def change
     create_table :strokes do |t|
-      t.float :velocity
-      t.float :angle
+      t.integer :velocity
+      t.integer :angle
       t.timestamps
     end
   end
